refactor: extract backup orchestration from App into BackupJob
App becomes a thin CLI shell (arg parsing, config load, Spark lifecycle, top-level error handling). BackupJob owns the workflow: enumerate source, run copy, log summary, fail-fast. Makes the pipeline testable in isolation.
